% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_segmentation.R
\name{matlab_old_segmentation}
\alias{matlab_old_segmentation}
\title{Coregister one or more source images to a reference image}
\usage{
matlab_old_segmentation(session, sources)
}
\arguments{
\item{session}{A \code{matlab_session}, already set up via \code{matlab_setup_spm()}.}

\item{sources}{Character vector of source mri image paths to segment.}
}
\value{
The \code{session}, invisibly, with one log entry per source file.
}
\description{
Coregister one or more source images to a reference image
}
